requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
future==0.18.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
serializers==0.2.4
sqlparse==0.4.3
